+++ src/asp/Camera/RPC_XML.cc
@@ -124,7 +124,15 @@ void asp::ImageXML::parse(xercesc::DOMElement* node) {
   } catch (...) {
     // generation_time may not exist  
   }
-  
+
+  image_descriptor = "";
+  try {
+    cast_xmlch(get_node<DOMElement>(node, "IMAGEDESCRIPTOR")->getTextContent(),
+                image_descriptor);
+  } catch (...) {
+    // image_descriptor may not exist  
+  }
+
   // Multispectral or panchromatic, or maybe MS1 or something else
   //if (band_id != "P" && band_id != "Multi" && band_id != "MS1") 
   //  vw_throw(ArgumentErr() << "Expecting BANDID in the XML file to be 'P', 'Multi', or 'MS1'.\n");